Remove dependency on go-ethereum
We only use it for hashing
