+++ shlib/go.mod
@@ -6,13 +6,13 @@ require (
 	github.com/ethereum/go-ethereum v1.10.8
 	github.com/google/go-cmp v0.5.6
 	github.com/pkg/errors v0.9.1
+	golang.org/x/crypto v0.0.0-20210506145944-38f3c27a63bf
 	gotest.tools v2.2.0+incompatible
 	gotest.tools/v3 v3.0.3
 )
 
 require (
 	github.com/btcsuite/btcd v0.22.0-beta // indirect
-	golang.org/x/crypto v0.0.0-20210506145944-38f3c27a63bf // indirect
 	golang.org/x/sys v0.0.0-20210816183151-1e6c022a8912 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
