+++ shlib/shcrypto/shuffle.go
@@ -4,14 +4,13 @@ import (
 	"encoding/binary"
 	"math/rand"
 
-	"github.com/ethereum/go-ethereum/crypto"
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
 )
 
 // computeSeed computes a seed value from the EpochSecretKey.
 func computeSeed(key *EpochSecretKey) int64 {
 	keyBytes := (*bn256.G1)(key).Marshal()
-	keyHash := crypto.Keccak256(keyBytes)
+	keyHash := keccak256(keyBytes)
 	return int64(binary.LittleEndian.Uint64(keyHash[:8]))
 }
